"Agrego cambios solicitados"

+++ core/src/main/Java/output/ProjectRepository.java
@@ -3,7 +3,7 @@ package output;
 import model.Project;
 
 public interface ProjectRepository {
-Project findProjectById(int id);
+Project findProjectById(Long id);
 boolean existsByName (String name);
 Project save (Project project);
 }
